requirements.txt
dash
pandas
mysql-connector-python
flask
plotly
seaborn
scikit-learn